requirements.txt
# requirements.txt
pytest==7.4.0
requests==2.31.0
PyYAML==6.0.1
allure-pytest==2.13.2
configparser==5.3.0
openpyxl==3.1.2